app/requirements.txt
requests
flask
redis
pystrich
psycopg2-binary
